LayoutTests/svg/filters/feBlend-invalid-mode-expected.txt

CONSOLE MESSAGE: TypeError: Type error
PASS: Invalid blend modes do not trigger a crash.
